fixes for web

+++ tests/e2e/hook.js
@@ -3,9 +3,6 @@ import { Application } from 'spectron';
 import electronPath from 'electron';
 import pathLib from 'path';
 
-const winMinio = pathLib.resolve(__dirname, '../bin/minio.exe');
-const unixMinio = 'minio';
-
 // Spectron API https://github.com/electron/spectron
 // Webdriver.io http://webdriver.io/api.html
 
@@ -30,12 +27,15 @@ export async function clearLocalStorage() {
     // await global.app.webContents.session.clearStorageData();
     global.app.webContents.reload();
   }*/
-  if (global.isWeb) {
-    localStorage.clear();
-    location.reload();
-  } else {
-    await global.app.webContents.executeJavaScript('localStorage.clear()');
-    global.app.webContents.reload();
+  if (!global.isUnitTest) {
+    if (global.isWeb) {
+      // global.client.executeScript('localStorage.clear()');
+      // global.client.clearLocalStorage();
+      global.client.refresh();
+    } else {
+      await global.app.webContents.executeJavaScript('localStorage.clear()');
+      global.app.webContents.reload();
+    }
   }
   // browser.clearLocalStorage();
   // global.app.client.localStorage('DELETE');
@@ -45,19 +45,37 @@ export async function clearLocalStorage() {
 export async function startSpectronApp() {
   if (global.isWeb) {
     const webdriverio = await require('webdriverio');
+    // https://webdriver.io/docs/configurationfile.html
     const options = {
       host: 'localhost', // Use localhost as chrome driver server
       port: 9515, // "9515" is the port opened by chrome driver.
       capabilities: {
-        browserName: 'chrome'
+        browserName: 'chrome',
         /*'goog:chromeOptions': {
           binary: electronPath, // Path to your Electron binary.
           args: [ /!* cli arguments *!/] // Optional, perhaps 'app=' + /path/to/your/app/
         }*/
+        timeouts: {
+          script: 60000
+        }
       },
+      waitforTimeout: 5000,
       logLevel: 'silent'
     };
+    // global.client = browser
     global.client = await webdriverio.remote(options);
+    // global.client.setTimeout({ 'script': 60000 });
+    /*client = await client
+      .init()
+      .setViewportSize({ width: 1024, height: 768 }, false)
+      .timeouts('script', 6000)
+      /!*
+                 Cannot set 'implicit' timeout because of a bug in webdriverio [1].
+                 [1] https://github.com/webdriverio/webdriverio/issues/974
+                 *!/
+      // .timeouts('implicit', 5000)
+      .timeouts('pageLoad', 30000);*/
+
     await global.client.url('http://localhost:8000');
   } else {
     global.app = new Application({
